Added a check which makes sure a winning submission can not be deleted

+++ app/Http/Controllers/ContestSubmissionController.php
@@ -109,6 +109,10 @@ class ContestSubmissionController extends Controller
     {
         $request->user()->can('delete', $submission);
 
+        if ($submission->winner) {
+            throw new Exception('You can not delete a winning submission');
+        }
+
         $submission->delete();
 
         return response('The submission has been deleted');
